fix integer overflow erorr when chunk is larger than 2^31 bytes.

+++ Converter/include/ConcurrentWriter.h
@@ -46,7 +46,7 @@ struct ConcurrentWriter {
 
 		this->tStart = now();
 
-		for (int i = 0; i < numThreads; i++) {
+		for (int64_t i = 0; i < numThreads; i++) {
 			threads.emplace_back([&]() {
 				flushThread();
 			});
